Переписал генерацию данных в контроллерах.
Для удобства тестирования данных, которые уходят в рендер, были добавлены вспомогательные функции в контроллеры.
