+++ README.md
@@ -86,3 +86,9 @@ npm run test:integration
 4. Содержание файлов:
     - возвращает корректные данные;
     - на рендер отправляются корректные аргументы.
+
+### Рефакторинг
+
+Для возможности написания модульных тестов выполнил рефакторинг некоторых участков кода.
+
+1. **Контроллеры**. Для тестирования данных, которые отправляются на рендер, были переписаны контроллеры (кажется, что так код выглядит понятней, контроллеры стали меньше). Для каждого контроллера написана отдельная функция в файле `generateData.js` папка utils.
