+++ controllers/contentController.js
@@ -1,5 +1,5 @@
 const { gitFileContent, gitFileTree } = require('../utils/git');
-const { buildFolderUrl, buildBreadcrumbs } = require('../utils/navigation');
+const { generateContent } = require('../utils/generateData');
 
 module.exports = function(req, res, next) {
   const { hash } = req.params;
@@ -14,11 +14,7 @@ module.exports = function(req, res, next) {
     .then(
       content => {
         if (content) {
-          res.render('content', {
-            title: 'content',
-            breadcrumbs: buildBreadcrumbs(hash, path.join('/')),
-            content
-          });
+          res.render('content', generateContent(hash, path, content));
         } else {
           next();
         }
