+++ controllers/filesController.js
@@ -1,20 +1,5 @@
 const { gitFileTree } = require('../utils/git');
-const {
-  buildFolderUrl,
-  buildFileUrl,
-  buildBreadcrumbs
-} = require('../utils/navigation');
-
-function buildObjectUrl(parentHash, { path, type }) {
-  switch (type) {
-    case 'tree':
-      return buildFolderUrl(parentHash, path);
-    case 'blob':
-      return buildFileUrl(parentHash, path);
-    default:
-      return '#';
-  }
-}
+const { generateFiles } = require('../utils/generateData');
 
 module.exports = function(req, res, next) {
   const { hash } = req.params;
@@ -24,17 +9,7 @@ module.exports = function(req, res, next) {
 
   return gitFileTree(hash, path).then(
     list => {
-      const files = list.map(item => ({
-        ...item,
-        href: buildObjectUrl(hash, item),
-        name: item.path.split('/').pop()
-      }));
-
-      res.render('files', {
-        title: 'files',
-        breadcrumbs: buildBreadcrumbs(hash, pathParam.join('/')),
-        files
-      });
+      res.render('files', generateFiles(hash, pathParam, list));
     },
     err => next(err)
   );
