+++ controllers/indexController.js
@@ -1,19 +1,11 @@
 const { gitHistory } = require('../utils/git');
 const { buildFolderUrl, buildBreadcrumbs } = require('../utils/navigation');
+const { generateList } = require('../utils/generateData');
 
 module.exports = function(req, res) {
   gitHistory(1, 20).then(
     history => {
-      const list = history.map(item => ({
-        ...item,
-        href: buildFolderUrl(item.hash, '')
-      }));
-
-      res.render('index', {
-        title: 'history',
-        breadcrumbs: buildBreadcrumbs(),
-        list
-      });
+      res.render('index', generateList(history));
     },
     err => next(err)
   );
